+++ app/src/main/java/com/shoppreglobal/shoppre/UI/Locker/LockerViewPackage.java
@@ -354,7 +354,7 @@ public class LockerViewPackage extends Fragment {
         packageId.setText("#" + String.valueOf(list.getId()));
         weight.setText(String.valueOf(list.getWeight()));
 
-        int sum = 0;
+        Double sum = Double.valueOf(0);
         for (int i = 0; i < list.getPackageItems().size(); i++) {
             sum = sum + list.getPackageItems().get(i).getPriceAmount();
         }
